Version 0.94b
A bug for the representation conversion, reported by Michal Kvasnica, was
fixed. The earlier 094* versions prematurely terminate the conversion when the
number of rows is equal to the number of columns in the input. This means the
earlier 094* do not compute correctly for simplices, for example.

Imported in Apr 2018 from cddlib-094b.tar.gz released Aug 25, 2005.