Version 0.94d
A bug of reporting a wrong unbounded direction vector of a dual inconsistent LP
is fixed. This bug occurs only if the exact GMP version of the dd_LPSolve is
used. For example, this error may occur in scdd_gmp. This error was reported by
Lars Schewe.

Imported in Apr 2018 from cddlib-094d.tar.gz released Feb 12, 2007.