Version 0.94e
A bug of reporting a wrong (sign) certificate of an infeasible LP is fixed.
This bug reported by Charles Geyer, occurs if the exact GMP version of the
dd_LPSolve is used with minimization. A bug of reporting infeasibility of a
feasible LP with column non-full rank LP is fixed. This bug was reported by
Sven Verdoolaege.

Imported in Apr 2018 from cddlib-094e.tar.gz released Jan 27, 2008.