Version 0.94i
The missing definition of dd_set_d function in the case of -DGMPRATIONAL flag
is used. The only change is in cddmp.h. The use of this function is explained
in testcdd2.c program. This function is useful when input is mostly float
(real). Even such a case, running on the rational exact arithmetic might be
helpful to debug the floating point computation.

Imported in Apr 2018 from cddlib-094i.tar.gz released Mar 9, 2018.