Properly detect GMP
Looking for "main" is cheating a bit (main is always there.) Instead we look
for __gmpz_init as proposed by GMP. If GMP can not be found, we don't build the
GMP enabled parts.

Also this drops obsolete checks (according to autoconf's manual) such as
AC_HEADER_STDC, AC_C_CONST (check that "const" is supported).

cddlib.so is now in theory versioned. We probably want to stick to 0.0.0 as
some people might rely on this?

Finally, this drops a bunch of hackish variables in the .am files around GMP.