configure.ac: use automake's "foreign" strictness level.
The default "strictness" level with automake is "gnu", which checks
that the GNU coding standards are being followed:

  https://www.gnu.org/software/automake/manual/html_node/Strictness.html

Since this is not a GNU project, that default is probably only an
annoyance. For example, with the level set to "foreign", we can
now delete the symlinks for the HISTORY, NEWS, and README files
which presumably exist only to avoid warnings from automake.