lib-src/Makefile{,.gmp}.am: use pkginclude directory for headers.
Using include_HEADERS in these two automake files results in the
headers being installed directly into the --includedir passed to
the ./configure script. This has caused various problems over the
years, leading distributions to override that decision and install
them in a subdirectory, either "cdd" (Debian) or "cddlib" (Fedora).
And of course Gentoo places them where the build system places them,
at the top level.

Automake allows you to use pkginclude_HEADERS instead, which places
the headers in the $(PACKAGE) subdirectory. In our case, that's going
to be "cddlib", because the package name is defined to be "cddlib" in
configure.ac. From here on out, consumers should look for our headers
in e.g. /usr/include/cddlib.

The decision to use "cddlib" instead of "cdd" was made for two reasons:

  1. Both locations have already been used on different distributions,
     so there is no backwards-compatible choice.

  2. The pkginclude_HEADERS support is built-in to autotools, and is
     therefore a tiny bit more standard.

Closes: https://github.com/cddlib/cddlib/issues/26