src/*.c: don't include headers from the system path.
This commit changes all of the C files under src/ to use

  #include "setoper.h"
  #include "cdd.h"

instead of

  #include <setoper.h>
  #include <cdd.h>

This is a weird case. These are "example" programs, so the obviously
correct thing to do is what you would do when writing a new program
that uses cddlib: assume that cddlib is installed, and use those
installed headers from the system. In light of recent changes, that
would look like

  #include <cddlib/setoper.h>
  #include <cddlib/cdd.h>

However, these examples are not currently installed as examples
(Github issue #24). They are built and installed as part of the
default cddlib build, and several of the programs are actually used by
other software. SageMath, for example, uses the scdd executable. And
when building these programs, it's important that they be linked
against the correct headers; namely the ones in the tarball, and not
the ones (perhaps from a previous version) already installed on the
system. The recent change to the header location highlighted this
issue, but it was already present in older releases.

So when building the src/ programs as part of cddlib, we want to
include the headers from lib-src/, which is what this commit does.
This kicks the can down the road; if we ever install these C files as
examples -- as documentation -- then we may need to tweak the #include
directives before installing them. The lib-src/Makefile.gmp.am file
already does something like this, so it's not without precedent.

Issue: https://github.com/cddlib/cddlib/issues/24